bug corrections

+++ reploy.rb
@@ -1,42 +1,56 @@
 require "rubygems"
-require "open4"
+#require "open4"
+
+repack_file_name = ''
 
 Dir['repack_*.tar'].each do |file_name|
-  cmd = "tar -xf " + file_name
-  #pid, stdin, showout, showerr = Open4::popen4 cmd
+  puts "Reploing file: " + file_name
+  tar_cmd = "tar -xf " + file_name
+  repack_file_name = file_name
+  system tar_cmd
+  #pid, stdin, showout, showerr = Open4::popen4 tar_cmd
   break
 end
 
 j = 0
 
-File.open('manifest.txt', 'r').each do |file_manifest|
-  if j == 0
-    j = j + 1
-    next
-  end
+# If it's a initial deploy, there is no manifest.txt file.
+unless repack_file_name == 'repack_initial.tar'
   
-  f = file_manifest.sub(/[\t]/, ',')
+  File.open('reploy_manifest.txt', 'r').each do |file_manifest|
+    if j == 0
+      j = j + 1
+      next
+    end
   
-  op, file = f.split(',')
+    f = file_manifest.sub(/[\t]/, ',')
   
-  case op
-    when 'A'
-      puts 'Add    : ' + file
-    when 'M'
-      puts 'Modify : ' + file
-    when 'D'
-      cmd = "rm " + file
-      pid, stdin, showout, showerr = Open4::popen4 cmd
+    op, file = f.split(',')
+  
+    case op
+      when 'A'
+        puts 'Add    : ' + file
+      when 'M'
+        puts 'Modify : ' + file
+      when 'D'
+        rm_cmd = "rm " + file
+        #pid, stdin, showout, showerr = Open4::popen4 cmd
+        system rm_cmd
       
-      puts 'Delete : ' + file
-  end
+        puts 'Delete : ' + file
+    end
   
-  j = j + 1
+    j = j + 1
+  end
+
 end
 
+rm_tar_cmd = "rm repack_*.tar"
+#pid, stdin, showout, showerr = Open4::popen4 rm_tar_cmd
+system rm_tar_cmd
 
-cmd = "rm repack_*.tar"
-#pid, stdin, showout, showerr = Open4::popen4 cmd
+# Delete reploy_manifest.txt
+system 'rm reploy_manifest.txt'
 
 puts "Reploy finished successful."
